Cardano.Client.Subscription

Documentation

subscribe ∷ RunNode blk ⇒ LocalSnocket → CodecConfig blk → NetworkMagic → NetworkClientSubcriptionTracers → ClientSubscriptionParams () → (NodeToClientVersion → ClientCodecs blk IO → ConnectionId LocalAddress → NodeToClientProtocols 'InitiatorMode ByteString IO x y) → IO Void

data MuxMode where

Constructors

InitiatorMode ∷ MuxMode

ResponderMode ∷ MuxMode

InitiatorResponderMode ∷ MuxMode

type ClientCodecs blk (m ∷ Type → Type) = Codecs' blk blk DeserialiseFailure m ByteString ByteString ByteString ByteString

data ConnectionId addr

Connection is identified by local and remote address.

TODO: the type variable which this data type fills in is called peerid. We should renamed to connectionId.

data LocalAddress

Local address, on Unix is associated with AF_UNIX family, on

Windows with `named-pipes`.

data NodeToClientProtocols (appType ∷ MuxMode) bytes (m ∷ Type → Type) a b

Recorod of node-to-client mini protocols.

Constructors

NodeToClientProtocols

Fields

localChainSyncProtocol ∷ RunMiniProtocol appType bytes m a b

local chain-sync mini-protocol

localTxSubmissionProtocol ∷ RunMiniProtocol appType bytes m a b

local tx-submission mini-protocol

localStateQueryProtocol ∷ RunMiniProtocol appType bytes m a b

local state-query mini-protocol

localTxMonitorProtocol ∷ RunMiniProtocol appType bytes m a b

local tx-monitor mini-protocol

type family BlockNodeToClientVersion blk

data MuxPeer bytes (m ∷ Type → Type) a where

Constructors

MuxPeer ∷ ∀ (pr ∷ PeerRole) ps (st ∷ ps) failure bytes (m ∷ Type → Type) a. (Show failure, ∀ (st' ∷ ps). Show (ClientHasAgency st'), ∀ (st' ∷ ps). Show (ServerHasAgency st'), ShowProxy ps) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → Peer ps pr st m a → MuxPeer bytes m a

MuxPeerPipelined ∷ ∀ (pr ∷ PeerRole) ps (st ∷ ps) failure bytes (m ∷ Type → Type) a. (Show failure, ∀ (st' ∷ ps). Show (ClientHasAgency st'), ∀ (st' ∷ ps). Show (ServerHasAgency st'), ShowProxy ps) ⇒ Tracer m (TraceSendRecv ps) → Codec ps failure m bytes → PeerPipelined ps pr st m a → MuxPeer bytes m a

MuxPeerRaw ∷ ∀ (m ∷ Type → Type) bytes a. (Channel m bytes → m (a, Maybe bytes)) → MuxPeer bytes m a

data MuxTrace

Enumeration of Mux events that can be traced.

data RunMiniProtocol (mode ∷ MuxMode) bytes (m ∷ Type → Type) a b where

Constructors

InitiatorProtocolOnly ∷ ∀ bytes (m ∷ Type → Type) a. MuxPeer bytes m a → RunMiniProtocol 'InitiatorMode bytes m a Void

ResponderProtocolOnly ∷ ∀ bytes (m ∷ Type → Type) b. MuxPeer bytes m b → RunMiniProtocol 'ResponderMode bytes m Void b

InitiatorAndResponderProtocol ∷ ∀ bytes (m ∷ Type → Type) a b. MuxPeer bytes m a → MuxPeer bytes m b → RunMiniProtocol 'InitiatorResponderMode bytes m a b

data WithMuxBearer peerid a

Type used for tracing mux events.

data ControlMessage

Control signal sent to a mini-protocol. expected to exit, on Continue it should continue its operation

Constructors

Continue
Continue operation.

Quiesce
Hold on, e.g. do not sent messages until resumed. This is not used for any hot protocol.

Terminate
The client is expected to terminate as soon as possible.

cChainSyncCodec ∷ Codecs' blk serialisedBlk e m bCS bTX bSQ bTM → Codec (ChainSync serialisedBlk (Point blk) (Tip blk)) e m bCS

cStateQueryCodec ∷ Codecs' blk serialisedBlk e m bCS bTX bSQ bTM → Codec (LocalStateQuery blk (Point blk) (Query blk)) e m bSQ

cTxSubmissionCodec ∷ Codecs' blk serialisedBlk e m bCS bTX bSQ bTM → Codec (LocalTxSubmission (GenTx blk) (ApplyTxErr blk)) e m bTX